refactor: 로그인에 좌우되는 프로필, 알람 박스를 하나로 통일
