+++ components/Header/GlobalNavBar/index.tsx
@@ -5,16 +5,15 @@ import {
   Logo,
   NavStyled,
   BoxStyled,
-  LoggedInBox,
+  ButtonBox,
 } from "./styled";
 import MobileMenuBox from "../MobileMenuBox";
 import { BOARD_LIST } from "../../../enum";
 import Button from "../../Button";
 import { useAppDispatch, useAppSelector } from "@/store/hooks";
 import { openModal } from "@/store/modalSlice";
-import Alarm from "../../Alarm";
-import { useRef, useState } from "react";
-import useOutsideClick from "@/hooks/common/useOutsideClick";
+
+import LoggedInMenuBox from "@/components/Header/LoggedInMenuBox";
 
 interface GlobalNavBarProps {
   logout: () => void;
@@ -25,15 +24,11 @@ export default function GlobalNavBar({
   logout,
   handleMobileMenu,
 }: GlobalNavBarProps) {
-  const [openAlarm, setOpenAlarm] = useState(false);
-  const AlarmListRef = useRef(null);
   const dispatch = useAppDispatch();
   const { token } = useAppSelector((state) => state.auth);
 
-  useOutsideClick(AlarmListRef, () => setOpenAlarm(false));
-
   return (
-    <Wrapper ref={AlarmListRef}>
+    <Wrapper>
       <HeaderStyled>
         <Logo>
           <Link href="/">사이드이펙트</Link>
@@ -46,27 +41,20 @@ export default function GlobalNavBar({
           ))}
         </NavStyled>
         <BoxStyled>
-          {token && (
-            <LoggedInBox>
-              <div>프로필</div>
-              <Alarm openAlarm={openAlarm} setOpenAlarm={setOpenAlarm} />
-            </LoggedInBox>
-          )}
-          {token ? (
-            <Button onClick={logout}>로그아웃</Button>
-          ) : (
-            <Button
-              onClick={() => dispatch(openModal({ modalType: "LoginModal" }))}
-            >
-              로그인
-            </Button>
-          )}
+          {token && <LoggedInMenuBox />}
+          <ButtonBox>
+            {token ? (
+              <Button onClick={logout}>로그아웃</Button>
+            ) : (
+              <Button
+                onClick={() => dispatch(openModal({ modalType: "LoginModal" }))}
+              >
+                로그인
+              </Button>
+            )}
+          </ButtonBox>
         </BoxStyled>
-        <MobileMenuBox
-          handleMobileMenu={handleMobileMenu}
-          openAlarm={openAlarm}
-          setOpenAlarm={setOpenAlarm}
-        />
+        <MobileMenuBox handleMobileMenu={handleMobileMenu} />
       </HeaderStyled>
     </Wrapper>
   );
