+++ components/Header/GlobalNavBar/styled.ts
@@ -47,27 +47,14 @@ export const NavStyled = styled.nav`
 `;
 
 export const BoxStyled = styled.div`
-  width: 200px;
-  background: aquamarine;
   display: flex;
-  justify-content: space-between;
   align-items: center;
-
-  ${mediaQuery("mobile")`
-    display: none;
-  `}
-
-  button {
-    svg {
-      transform: scale(1.2);
-    }
-  }
+  justify-content: flex-end;
+  gap: 1rem;
 `;
 
-export const LoggedInBox = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  margin-right: 1rem;
+export const ButtonBox = styled.div`
+  ${mediaQuery("mobile")`
+    display: none;
+  `};
 `;
