+++ components/Header/MobileMenuBox/styled.ts
@@ -7,7 +7,6 @@ export const Wrapper = styled.div`
     display: flex;
     justify-contents: center;
     align-items: center;
-    gap: 1rem;
   `}
 
   display: none;
